requirements.txt
absl-py==0.1.12
appnope==0.1.0
astor==0.6.2
attrs==17.4.0
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cycler==0.10.0
DateTime==4.2
decorator==4.1.2
descartes==1.1.0
entrypoints==0.2.3
facebook-sdk==2.0.0
Fiona==1.7.11.post1
gast==0.2.0
geopandas==0.3.0
geopy==1.12.0
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
ijson==2.3
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
munch==2.2.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.2.2
numpy==1.13.3
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
populartimes==2.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
py==1.5.2
pycurl==7.43.0.1
Pygments==2.2.0
pyparsing==2.2.0
pyproj==1.9.5.1
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
statsmodels==0.8.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
zope.interface==4.4.3
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
bokeh==0.12.14
certifi==2018.1.18
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cycler==0.10.0
decorator==4.2.1
descartes==1.1.0
entrypoints==0.2.3
Fiona==1.7.11.post1
flexx==0.4.1
geopandas==0.3.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
munch==2.2.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
oauthlib==2.0.7
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
pyproj==1.9.5.1
PySocks==1.6.8
python-dateutil==2.7.0
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
tweepy==3.6.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
